pyproject.toml
[project]
name = "chroma-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "python-frontmatter>=1.1.0",
    "tokenizers>=0.22.1",
]

[dependency-groups]
dev = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "markdownify>=1.2.2",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "requests>=2.32.5",
]


[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
]
